GreenRoute AI/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
tensorflow==2.20.0
streamlit==1.54.0
streamlit-autorefresh==1.0.1
scikit-learn==1.8.0
requests==2.32.5
python-multipart==0.0.22
pydantic==2.12.5
numpy==2.4.0
packaging==25.0
multitasking==0.0.12
keras==3.13.2
pandas==2.3.3
pillow==10.4.0